code2docs-cli/pyproject.toml
[tool.poetry]
name = "code2docs-cli"
version = "0.1.1"
description = ""
authors = ["xKarinSan <setsunayukii88@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
code2docs-cli = "code2docs_cli.main:app"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "1.1.0"
typer = "0.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
